requirements.txt
python-telegram-bot==20.3
httpx==0.24.1
tenacity==8.2.2
python-dotenv==1.0.0